move app component to routes
