+++ frontend/src/routes/App.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { useState, useEffect } from "react";
 import "./App.css";
-import Button from "./components/button";
-import Navbar from "./components/nav";
+import Button from "../components/button";
+import Navbar from "../components/nav";
 
 const App: React.FC = () => {
   let [counter, setCounter] = useState(0);
